Add _xarray module
